pyproject.toml
[tool.poetry]
name = "jhapa"
version = "0.1.0"
description = "A simple link aggregator"
authors = ["keshab <self@keshab.net>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
django-mptt = "^0.10.0"
bleach = "^3.1"
Pillow = "^6.2"
psycopg2 = "^2.8"

[tool.poetry.dev-dependencies]
ipython = "^7.8"
pylint = "^2.4"
django-debug-toolbar = "^2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
